fix: 회원가입 후 쿠폰 발급 및 포인트 적립 비동기 처리 에러 해결
