+++ src/main/java/shop/wannab/userservice/user/service/PostSignupService.java
@@ -1,0 +1,9 @@
+package shop.wannab.userservice.user.service;
+
+import org.springframework.scheduling.annotation.Async;
+import shop.wannab.userservice.user.domain.entity.User;
+
+public interface PostSignupService {
+    @Async
+    void handlePostSignup(User user);
+}
